+++ Changelog.md
@@ -6,6 +6,7 @@
 * added plus-i-state and minus-i-state functions
 * added |+i⟩ and |-i⟩ states
 * use matrix instead of phase gate to define s and s-dag gates
+* moved oracle functions from grover-test to grover ns
 * enhanced tutorial
 
 ## Version 0.4.0